% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegDesterieux-package.R
\docType{package}
\name{ggsegDesterieux-package}
\alias{ggsegDesterieux}
\alias{ggsegDesterieux-package}
\title{ggsegDesterieux: Destrieux Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Destrieux cortical atlas (aparc.a2009s) for the 'ggseg' ecosystem. Provides a 'ggseg_atlas' object with 2D polygon geometry for use with 'ggseg'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegDesterieux}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegDesterieux/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
